backend/app/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
psycopg2-binary
requests
python-dotenv
pymupdf
google-generativeai
assemblyai
yt-dlp
